Fix two potential integer overflows
Found by CodeQL

+++ src/crgContactPoint.c
@@ -456,7 +456,7 @@ crgContactPointPreloadHistoryU( CrgContactPointStruct *cp, double u )
 
     /* --- shift previous values --- */
     if ( cp->history.totalSize > 1 )
-        memmove( &( cp->history.entry[1] ), cp->history.entry, ( cp->history.totalSize - 1 ) * cp->history.entrySize );
+        memmove( &( cp->history.entry[1] ), cp->history.entry, (size_t)( cp->history.totalSize - 1 ) * cp->history.entrySize );
 
     /* --- compute x and y for u on reference line --- */
     frac = ( u - cp->crgData->channelU.info.first ) / cp->crgData->channelU.info.inc;
